Extract finishInit method to new module
